# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DropTracker Web API",
    "version": "1.0.0",
    "description": "Versioned, browser-oriented JSON API consumed by the Next.js BFF. Separate from the RuneLite intake API. See FRONTEND_PLAN.md §6."
  },
  "servers": [{ "url": "/api/v1" }],
...
  "paths": {
...
    "/admin/logs": {
      "get": {
        "tags": ["admin"],
        "operationId": "adminLogs",
        "summary": "Tail the application logs table",
        "security": [{ "session": [] }],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 500, "default": 200 }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent log entries and known sources.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AdminLogsResponse" }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dt_session"
      }
    },
...
    "responses": {
      "Error": {
        "description": "RFC-7807 problem details.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "RFC-7807 error envelope.",
        "required": ["title", "status"],
        "properties": {
          "type": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" }
        }
      },
...
      "AdminLogEntry": {
        "type": "object",
        "required": ["ts", "level", "source", "message"],
        "properties": {
          "ts": { "type": "integer", "description": "Unix seconds." },
          "level": { "type": "string" },
          "source": { "type": "string" },
          "message": { "type": "string" }
        }
      },
      "AdminLogsResponse": {
        "type": "object",
        "required": ["entries", "sources"],
        "properties": {
          "entries": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AdminLogEntry" }
          },
          "sources": { "type": "array", "items": { "type": "string" } }
        }
      },
...
    }
  }
}